xopen {xopen}	R Documentation

Open a file, directory or URL

Description

Open a file, directory or URL, using the local platforms conventions, i.e. associated applications, default programs, etc. This is usually equivalent to double-clicking on the file in the GUI.

Usage

xopen(target = NULL, app = NULL, quiet = FALSE, ...)

Arguments

target
String, the path or URL to open.

app
Specify the app to open target with, and its arguments, in a character vector. Note that app names are platform dependent.

quiet
Whether to echo the command to the screen, before running it.

...
Additional arguments, not used currently.

Examples

xopen("test.R")
xopen("https://ps.r-lib.org")
xopen(tempdir())

[Package xopen version 1.0.0 Index]